+++ classes/QuicksortRecursivo.h
@@ -10,8 +10,8 @@ using namespace std;
 class QuicksortRecursivo
 {
 private:
-long long int comparacoes;
-long long int trocas;
+unsigned long int comparacoes;
+unsigned long int trocas;
 ifstream leitor;
 
 public:
